fix: fixed issue with override existing alarm when creating new one
- updated AlarmDao.java, AlarmViewModel.java and AlarmViewModelFactory.java
- updated ResultListener.java
- updated Alarm.java
- updated MainActivity.java and NewAlarmActivity.java
- updated AlarmUtils.java

+++ app/src/main/java/vkurman/shiftedalarmclock/aac/AlarmDao.java
@@ -36,7 +36,7 @@ import vkurman.shiftedalarmclock.models.Alarm;
 public interface AlarmDao {
 
     @Query("SELECT * FROM alarm WHERE id = :alarmId")
-    LiveData<Alarm> load(int alarmId);
+    LiveData<Alarm> load(long alarmId);
 
     @Query("SELECT * FROM alarm")
     LiveData<List<Alarm>> loadAll();
